requirements.txt
Django==5.0.2
django-environ==0.11.2
mssql-django==1.3
pyodbc==5.0.1
qrcode==7.4.2
Pillow==9.5.0
django-crispy-forms==2.1
crispy-bootstrap4==2024.1
django-filter==23.5
reportlab==4.1.0
django-cors-headers==4.3.1
python-dotenv==1.0.1 